requirements.txt
flask==1.1.2
python-dotenv==0.14.0
celery==5.0.0
peewee==3.13.3
PyGithub==1.53
polib==1.1.0
redis==3.5.3